Fragmenting sent handshake messages
Previously handshake messages being sent out would not be fragmented.
Most handshake messages are pretty small but a couple of exceptions,
such as Certificate, mean the packet size can be greater than the MTU,
which then causes some packet loss and the handshake would never be
able to complete. This changes the Marshal interface to return an array
of packets, fragmented at a specified length so that the packets do not
go over the MTU.